requirements.txt
astropy~=5.3.1
numpy~=1.25.0
Pillow~=10.0.0
PyExifTool~=0.5.5
tqdm~=4.65
matplotlib~=3.7.2
rawpy~=0.18.1